refixing Contributor Table by directly inserting data into it from data.py, and API to fetch relevant data from the Table
