+++ Backend/data.py
@@ -211,3 +211,16 @@ challenges = [
     #     'challengeHintThree': 'PlaceHolder#3'
     # }
 ]
+
+Contributors = [
+    {
+        'challengeCode' : 'blockchain-cheap-glitch',
+        'Name' : 'Kiinzu',
+        'Link' : 'https://google.com'
+    },
+    {
+        'challengeCode' : 'blockchain-bar',
+        'Name' : 'Mewzael',
+        'Link' : 'https://bing.com'
+    }
+]
